requirements.txt
numpy
matplotlib
jupyter
opencv-python
pillow
scikit-learn
scikit-image
scipy
h5py   
gym
pyglet
pyopengl
pyvirtualdisplay
tensorflow-gpu==2.2
keras==2.2.3
tornado==5.0
msgpack-rpc-python
airsim
stable-baselines3[extra]
